Split nvhttp endpoint modules
