+++ cmake/compile_definitions/common.cmake
@@ -98,8 +98,16 @@ set(SUNSHINE_TARGET_FILES
         "${CMAKE_SOURCE_DIR}/src/webhook.h"
         "${CMAKE_SOURCE_DIR}/src/nvhttp.cpp"
         "${CMAKE_SOURCE_DIR}/src/nvhttp.h"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/abr_api.cpp"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/abr_api.h"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/apps.cpp"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/apps.h"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/display_control.cpp"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/display_control.h"
         "${CMAKE_SOURCE_DIR}/src/nvhttp/display_scale.cpp"
         "${CMAKE_SOURCE_DIR}/src/nvhttp/display_scale.h"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/dynamic_params.cpp"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/dynamic_params.h"
         "${CMAKE_SOURCE_DIR}/src/nvhttp/url_utils.cpp"
         "${CMAKE_SOURCE_DIR}/src/nvhttp/url_utils.h"
         "${CMAKE_SOURCE_DIR}/src/nvhttp_stream_start.cpp"
